+++ api/CHANGELOG.md
@@ -10,6 +10,8 @@ All notable changes to the **Prowler API** are documented in this file.
 ### Changed
 - Renamed field encrypted_password to password for M365 provider [(#7784)](https://github.com/prowler-cloud/prowler/pull/7784)
 
+### Fixed
+- Fixed Kubernetes UID validation to allow valid context names [(#XXXX)](https://github.com/prowler-cloud/prowler/pull/XXXX)
 ---
 
 ## [v1.8.2] (Prowler v5.7.2)